[charts-pro] Use new sequential color palette in FunnelChart
