+++ packages/x-charts/src/colorPalettes/index.ts
@@ -1,1 +1,10 @@
-export * from './colorPalettes';
+export * from './types';
+
+// CATEGORICAL PALETTES
+export * from './categorical/rainbowSurge';
+export * from './categorical/blueberryTwilight';
+export * from './categorical/mangoFusion';
+export * from './categorical/cheerfulFiesta';
+
+// SEQUENTIAL PALETTES
+export * from './sequential/strawberrySky';
