+++ packages/x-charts/src/colorPalettes/types.ts
@@ -1,0 +1,4 @@
+export type ChartsColorPaletteCallback = (mode: 'light' | 'dark') => string[];
+export type ChartsColorPalette = string[] | ChartsColorPaletteCallback;
+export type ChartsColorCallback = (mode: 'light' | 'dark') => string;
+export type ChartsColor = string | ChartsColorCallback;
